Fixes #11
This commit removes the cigarettes per day -> smoking status level -> initial COHB in blood and goes directly from cigarettes per day to initial COHB in blood without using the smoking status level as an intermediate step.  As a consequence the standard deviation for this parameter is not dependent upon the variation of number of cigarette smoked per day rather than a set value of 0.32*RSD.  Of course the old method of determining smoking status directly and using the 0.32*RSD is still available as an option that is used when uploading legacy PRN files but it is anticipated that this will be depreciated.